refactor(farm-group): split components
